refactor: inline finykStats.js passthroughs into domain/selectors
finykStats.js was a thin facade — 12 of 13 exports were pure re-exports
from `domain/selectors` (+ one alias + one unused helper). The only real
function, `getMonthlySpendSeries`, moves into selectors.ts (properly
typed) next to the `getMonthlySummary` it composes.

- Move getMonthlySpendSeries into domain/selectors.ts (typed).
- Redirect the single importer (useAnalytics.js) to domain/selectors.
- Delete src/modules/finyk/lib/finykStats.js.
- Drop the finykStats mention from README's lib/ inventory.
